Add RSpec
Moves to RSpec for the test suite. This installs rspec rails using the
`rails generate rspec:install` command. I also added configuration for
rspec to filter for the `focus` option if found. The default rake task
is now to run rspec, but you can run the specs with `rspec` as well.