add paperclip and file uploading
Adds the paperclip gem and configuration for AWS uploading, and adds the
attachments to the Campaign and Organization models. I removed the
logo_file_name field on the organizations table so that we could use
Paperclip's add_attachment migration method without conflicting with
column names.

This also adds the bucket name, and region to .env. I set the aws keys
in .env.local and set them on heroku, but wanted to keep those out of
version control. Development images are also ignored from version
control.