Add paperclip with s3 configuration (#39)
* add paperclip and file uploading

Adds the paperclip gem and configuration for AWS uploading, and adds the
attachments to the Campaign and Organization models. I removed the
logo_file_name field on the organizations table so that we could use
Paperclip's add_attachment migration method without conflicting with
column names.

This also adds the bucket name, and region to .env. I set the aws keys
in .env.local and set them on heroku, but wanted to keep those out of
version control. Development images are also ignored from version
control.

* add example scaffold_controller for organizations

Adds an organizations controller and views to test out the image
uploading. Before merging this commit should be reverted.

* remove test images

* Revert "add example scaffold_controller for organizations"

This reverts commit 4f848d6e2d25790bffbb3243ad66a13613beb29b. We don't
need these scaffolded out views and controllers for testing anymore.