add membership and remove donors (#46)
* add membership and remove donors

Add a membership model that belongs to an organization and a user, and
removes the donor model. The membership model not represents the concept
of a donor, so the model and table are no longer necessary.

* add email subscribe boolean to membership

Adds a boolean field on memberships to unsubscribe from a given
membership email list.