setup actionmailer in dev and prod
sets up actionmailer for use in development using gmail and production
using the sendgrid addon https://devcenter.heroku.com/articles/sendgrid

The devise mailer setting is just the email address that is displayed
when the email is sent, it doesn't have to exist.

If you want to use the mailer in development, you can place your creds
in a .env.local file in the format of

```
export GMAIL_USERNAME=username@gmail.com
export GMAIL_PASSWORD=mYs3cUr3pAsSw0rD
```

You may have to change some security settings in gmail, so it may be
best to set up another account for this.