Rubocop Cleanup (#108)
* correct style errors using rubocop autocorrect

I wanted to make sure our code was clean, so I ran rubocop with
default cops, auto generated the config, and autocrorrected the
errors that rubocop could autocorrect. I plan on correcting the rest
of the issues in follow up commits.

* fix all non-metrics, non-autocorrect-able rubocop warnings

the last commit was fixing all autocorrect rubocop issues
This commit is all of the offenses that were excluded in the todo but
are not metrics such as line, method, and class lengths.

* Set line length metric to 100 characters

Set the line length metric for rubocop to 100 characters and fix all
warnings

* Fix rubocop ABC size warnings to use rubocop default

* Change PGN => PNG & fix frozen string literal warning

* When using rvm + rubocop, 100+ warnings for a missing frozen string
literal comment at the top of the files. Rather than add the comment
and impact 100+ files, I've disabled the check