fix tests that broke after introducing uniqueness validation
A bunch of tests broke after introducing the uniqueness validation on
an organization name. This validation is necessary because of the slug
generated to be used for the friendly id.

This fixes those tests by making sure the associated organization has
a unique name.

It also adds back the rspec filter run that was removed